requirements.txt
alembic==1.10.2
click==8.0.4
email-validator==1.3.1
Flask==2.2.3
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==20.1.0
importlib-metadata==6.1.0
itsdangerous==2.1.1
Jinja2==3.0.3
Mako==1.2.4
MarkupSafe==2.1.1
psycopg2==2.9.5
PyJWT==2.6.0
python-dotenv==1.0.0
SQLAlchemy==2.0.8
typing_extensions==4.5.0
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
